+++ lab5/src/main/java/inno/tech/entity/RefProductClass.java
@@ -4,8 +4,6 @@ import jakarta.persistence.Entity;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
-import jakarta.persistence.JoinColumn;
-import jakarta.persistence.ManyToOne;
 import jakarta.persistence.OneToMany;
 import lombok.Getter;
 
@@ -17,8 +15,8 @@ public class RefProductClass {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long internalId;
-//    private String value;
+    private Integer internalId;
+    //    private String value;
     private String gbiCode;
     private String gbiName;
     private String productRowCode;
