+++ lab5/src/main/java/inno/tech/repositories/AccountPoolRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface AccountPoolRepository extends JpaRepository<AccountPool, Long> {
+public interface AccountPoolRepository extends JpaRepository<AccountPool, Integer> {
     Optional<AccountPool> findByBranchCodeAndCurrencyCodeAndMdmCodeAndPriorityCodeAndRegistryTypeCode(String branchCode, String currencyCode, String mdmCode, String priorityCode, String registryTypeCode);
 }
