+++ lab5/src/main/java/inno/tech/repositories/ProductRegisterRepository.java
@@ -5,8 +5,8 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface ProductRegisterRepository extends JpaRepository<ProductRegister, Long> {
+public interface ProductRegisterRepository extends JpaRepository<ProductRegister, Integer> {
 
-    Optional<ProductRegister> findByIdAndType(Long id, String type);
+    Optional<ProductRegister> findByIdAndType(Integer id, String type);
 }
