+++ lab5/src/main/java/inno/tech/repositories/RefProductClassRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface RefProductClassRepository extends JpaRepository<RefProductClass,Long> {
+public interface RefProductClassRepository extends JpaRepository<RefProductClass,Integer> {
     Optional<RefProductClass> findByValue(String number);
 }
